Traktor: Static code analysiz fixes. VS 2015 project files.
[git-p4: depot-paths = "//depot/traktor/": change = 17956]

+++ code/Core/Serialization/CompactSerializer.cpp
@@ -803,7 +803,7 @@ void CompactSerializer::operator >> (const Member< ISerializable* >& m)
 			typeId != 0x1f
 		)
 		{
-			if (typeId - 1 >= m_ntypes)
+			if (uint32_t(typeId - 1) >= m_ntypes)
 				return;
 
 			const TypeInfo* type = m_types[typeId - 1];
